requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
Babel==2.10.3
billiard==3.6.4.0
Brotli==1.0.9
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.4
crispy-bootstrap5==0.6
cssselect2==0.7.0
Deprecated==1.2.13
dj-database-url==1.0.0
Django==4.0.8
django-bootstrap-icons==0.8.1
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-countries==7.3.2
django-crispy-forms==1.14.0
django-environ==0.9.0
django-jet-reboot==1.3.1
django-paypal==2.0
django-phonenumber-field==7.0.0
django-timezone-field==5.0
feedparser==6.0.10
flower==1.2.0
fonttools==4.37.3
google-api-python-client==1.4.1
gunicorn==20.1.0
html5lib==1.1
httplib2==0.20.4
humanize==4.4.0
idna==3.3
kombu==5.2.4
mysqlclient==2.1.1
oauth2client==4.1.3
packaging==21.3
phonenumbers==8.12.55
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psycopg2==2.9.3
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.3.0
pyparsing==3.0.9
pyphen==0.13.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
redis==4.3.4
requests==2.28.1
rsa==4.9
sgmllib3k==1.0.0
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
whitenoise==6.2.0
wrapt==1.14.1
zopfli==0.2.1
